fix(macros): passing on crate root
The 'crate_root' attribute ensures that rasn macros generate valid code even if rasn's types and traits are imported from another origin than 'rasn'. A typical example are rasn's own unit tests, which import from 'crate' instead of 'rasn'.
This commit fixes an error, where the 'crate_root' attribute is not passed on to opaque types that rasn creates internally to represent set fields. Because of this error, structs that were annotated with 'crate_root' and 'set' still contained references to the 'rasn' crate, which led to compiler errors.